+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "clairvoyant" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 3.1.0 - 2019-11-27
+
+### Added
+
+- Enhanced log output.
+- `clairvoyant.outputChannelVolume` setting.
+
 ## 3.0.1 - 2019-11-26
 
 ### Fixed